phase0(infra): fix clang-format and clang-tidy warnings
- Untrack build/ from git; expand .gitignore to cover all common artifacts
- CMakeLists.txt: tidy target strips -fcontracts via sh -c + sed before
  running clang-tidy (clang doesn't understand GCC-only -fcontracts)
- .clang-tidy: add LocalConstantCase/StaticConstantCase lower_case,
  disable cppcoreguidelines-macro-usage and modernize-use-std-print
- primitives.hpp: add ~Primitive() = default; rename tag structs to CamelCase
  (I8Tag, U8Tag, ...) to satisfy readability-identifier-naming
- error.hpp: use uint8_t enum base (performance-enum-size); add <cstdint>
- log.hpp: use uint8_t enum base, add braces to all if statements,
  replace fprintf with std::println (cppcoreguidelines-pro-type-vararg)
- error_test.cpp: replace C-style array with std::array (modernize-avoid-c-arrays)

+++ src/utils/error.hpp
@@ -14,6 +14,7 @@
 // Thread-safety: EdbError is a plain enum; EdbResult<T> follows the
 // thread-safety of std::expected (value-type semantics, no shared state).
 
+#include <cstdint>
 #include <expected>
 #include <string_view>
 
@@ -22,25 +23,25 @@ namespace edb {
 // ---------------------------------------------------------------------------
 // Error codes
 // ---------------------------------------------------------------------------
-enum class EdbError : int {
+enum class EdbError : uint8_t {  // raw-primitive: enum base type requires stdint typedef
     // Generic
     Ok = 0,           // not an error; used as a sentinel
     InvalidArgument,  // precondition violated by caller
     OutOfMemory,
-    NotSupported,     // operation not implemented by this backend
-    NotFound,         // requested object does not exist
+    NotSupported,  // operation not implemented by this backend
+    NotFound,      // requested object does not exist
     AlreadyExists,
-    Overflow,         // arithmetic or buffer overflow
-    Corruption,       // on-disk data failed integrity check (checksum etc.)
+    Overflow,    // arithmetic or buffer overflow
+    Corruption,  // on-disk data failed integrity check (checksum etc.)
 
     // I/O
-    IoError,          // underlying read/write failed
+    IoError,  // underlying read/write failed
     IoTimeout,
-    IoAlignment,      // buffer or offset not aligned to sector size (xNVMe)
+    IoAlignment,  // buffer or offset not aligned to sector size (xNVMe)
 
     // Storage engine
     PageNotFound,
-    BufferPoolFull,   // all frames pinned; cannot evict
+    BufferPoolFull,  // all frames pinned; cannot evict
     InvalidPageId,
 
     // Transactions
@@ -73,29 +74,52 @@ using EdbStatus = EdbResult<void>;
 // ---------------------------------------------------------------------------
 constexpr std::string_view edb_error_name(EdbError e) noexcept {
     switch (e) {
-        case EdbError::Ok:                    return "Ok";
-        case EdbError::InvalidArgument:       return "InvalidArgument";
-        case EdbError::OutOfMemory:           return "OutOfMemory";
-        case EdbError::NotSupported:          return "NotSupported";
-        case EdbError::NotFound:              return "NotFound";
-        case EdbError::AlreadyExists:         return "AlreadyExists";
-        case EdbError::Overflow:              return "Overflow";
-        case EdbError::Corruption:            return "Corruption";
-        case EdbError::IoError:               return "IoError";
-        case EdbError::IoTimeout:             return "IoTimeout";
-        case EdbError::IoAlignment:           return "IoAlignment";
-        case EdbError::PageNotFound:          return "PageNotFound";
-        case EdbError::BufferPoolFull:        return "BufferPoolFull";
-        case EdbError::InvalidPageId:         return "InvalidPageId";
-        case EdbError::TransactionAborted:    return "TransactionAborted";
-        case EdbError::DeadlockDetected:      return "DeadlockDetected";
-        case EdbError::SnapshotTooOld:        return "SnapshotTooOld";
-        case EdbError::TypeNotFound:          return "TypeNotFound";
-        case EdbError::TypeAlreadyRegistered: return "TypeAlreadyRegistered";
-        case EdbError::CatalogCorrupted:      return "CatalogCorrupted";
-        case EdbError::ParseError:            return "ParseError";
-        case EdbError::AnalyzerError:         return "AnalyzerError";
-        case EdbError::ExecutorError:         return "ExecutorError";
+        case EdbError::Ok:
+            return "Ok";
+        case EdbError::InvalidArgument:
+            return "InvalidArgument";
+        case EdbError::OutOfMemory:
+            return "OutOfMemory";
+        case EdbError::NotSupported:
+            return "NotSupported";
+        case EdbError::NotFound:
+            return "NotFound";
+        case EdbError::AlreadyExists:
+            return "AlreadyExists";
+        case EdbError::Overflow:
+            return "Overflow";
+        case EdbError::Corruption:
+            return "Corruption";
+        case EdbError::IoError:
+            return "IoError";
+        case EdbError::IoTimeout:
+            return "IoTimeout";
+        case EdbError::IoAlignment:
+            return "IoAlignment";
+        case EdbError::PageNotFound:
+            return "PageNotFound";
+        case EdbError::BufferPoolFull:
+            return "BufferPoolFull";
+        case EdbError::InvalidPageId:
+            return "InvalidPageId";
+        case EdbError::TransactionAborted:
+            return "TransactionAborted";
+        case EdbError::DeadlockDetected:
+            return "DeadlockDetected";
+        case EdbError::SnapshotTooOld:
+            return "SnapshotTooOld";
+        case EdbError::TypeNotFound:
+            return "TypeNotFound";
+        case EdbError::TypeAlreadyRegistered:
+            return "TypeAlreadyRegistered";
+        case EdbError::CatalogCorrupted:
+            return "CatalogCorrupted";
+        case EdbError::ParseError:
+            return "ParseError";
+        case EdbError::AnalyzerError:
+            return "AnalyzerError";
+        case EdbError::ExecutorError:
+            return "ExecutorError";
     }
     return "UnknownError";
 }
